requirements.txt
alabaster==0.7.12
alembic==1.7.3
anytree==2.8.0
asgiref==3.4.1
attrs==21.2.0
autopage==0.4.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
boto3==1.18.37
botocore==1.21.37
bullet==2.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
click==8.0.1
cliff==3.9.0
cmd2==2.2.0
colorama==0.4.4
configobj==5.0.6
configparser==5.0.2
cryptography==35.0.0
dnspython==2.0.0
docutils==0.17.1
ecdsa==0.17.0
email-validator==1.1.3
et-xmlfile==1.1.0
fastapi==0.68.1
gnupg==2.3.1
greenlet==1.1.1
h11==0.12.0
httptools==0.2.0
humanfriendly==9.2
idna==3.2
imagesize==1.2.0
ipwhois==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
lxml==4.6.3
Mako==1.1.5
MarkupSafe==2.0.1
numpy==1.21.2
openpyxl==3.0.9
packaging==21.0
pandas==1.3.3
passlib==1.7.4
pbr==5.6.0
pexpect==4.8.0
prettytable==2.2.0
property-manager==3.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
Pygments==2.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
python-secrets==21.9.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
setuptools-scm==6.3.2
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.15
starlette==0.14.2
stevedore==3.4.0
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.16.0
verboselogs==1.7
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
xkcdpass==1.19.3
yamlreader==3.0.4
